Update schemas to match the Paris edition of the protocol
This effectively means stepping up ArtP and CD to v3.1.0, but only
the former actually affects the schemas.

Some of the generated files that were present prior to this commit
appear to have been modified after they were generated, i.e.
regenerating them introduced various non-consequential changes
(whitespace and/or line endings).

The copyright header in each file was retained though, although we
should update the code that generates the files to keep the header.